+++ app/Livewire/Client/CourseCreation/Components/Builders/Lesson/LessonCreate.php
@@ -3,23 +3,22 @@
 namespace App\Livewire\Client\CourseCreation\Components\Builders\Lesson;
 
 use App\Traits\HasErrors;
-use App\Traits\WithLessonForm;
 use App\Validator\NewLessonValidator;
 use Exception;
 use Illuminate\Contracts\View\Factory;
 use Illuminate\Foundation\Application;
-use Illuminate\Support\Facades\File;
-use Illuminate\Support\Facades\Storage;
 use Illuminate\View\View;
+use Livewire\Attributes\Modelable;
 use Livewire\Attributes\On;
 use Livewire\Attributes\Validate;
 use Livewire\Component;
 
 class LessonCreate extends Component {
-    use HasErrors, WithLessonForm;
+    use HasErrors;
 
     #[Validate]
-    public array $lesson = [
+    #[Modelable]
+    public array $newLesson = [
         'title' => '',
         'video_file_name' => '',
         'document' => '',
@@ -30,26 +29,46 @@ class LessonCreate extends Component {
         'practice_assessments' => []
     ];
 
-    public string|int $moduleIndex;
-
     public array $existingLessonTitles = [];
+
+    public array $messages;
 
     public bool $assessmentValid = false;
 
     public function mount(): void
     {
-        $this->mountLessonForm();
+        $this->messages = NewLessonValidator::$MESSAGES;
+    }
+
+    public function rules(): array
+    {
+        $rules = NewLessonValidator::rules();
+
+        if (is_string($rules['newLesson.title'])) {
+            $rules['newLesson.title'] = explode('|', $rules['newLesson.title']);
+        }
+
+        $rules['newLesson.title'][] = function ($attribute, $value, $fail) {
+            $incomingTitle = mb_strtolower(trim((string)$value));
+
+            if ($incomingTitle !== '') {
+                $normalizedExisting = array_map(
+                    fn($t) => mb_strtolower(trim((string)$t)),
+                    $this->existingLessonTitles ?? []
+                );
+
+                if (in_array($incomingTitle, $normalizedExisting, true)) {
+                    $fail('Lesson title must be unique within the module.');
+                }
+            }
+        };
+
+        return $rules;
     }
 
     public function updated(): void
     {
         $this->validate();
-    }
-
-    #[On('create-lesson')]
-    public function receiveData(string|int $moduleIndex): void
-    {
-        $this->moduleIndex = $moduleIndex;
     }
 
     /**
@@ -58,17 +77,17 @@ class LessonCreate extends Component {
     public function addLesson(): void
     {
         try {
-            if (!empty($this->lesson['tmp_video_file_name'] ?? null) && empty($this->lesson['video_file_name'] ?? null)) {
-                $this->messages['lesson.video_file_name.required_if'] = 'You have a temporary uploaded video. Please save it before adding the lesson.';
+            if (!empty($this->newLesson['tmp_video_file_name'] ?? null) && empty($this->newLesson['video_file_name'] ?? null)) {
+                $this->messages['newLesson.video_file_name.required_if'] = 'You have a temporary uploaded video. Please save it before adding the lesson.';
             }
             $this->validate();
 
             if (
-                ($this->lesson['type'] ?? null) === 'assessment' &&
-                ($this->lesson['assessment']['type'] ?? null) === 'quiz' &&
+                ($this->newLesson['type'] ?? null) === 'assessment' &&
+                ($this->newLesson['assessment']['type'] ?? null) === 'quiz' &&
                 $this->assessmentValid === false
             ) {
-                $this->addError('lesson.assessment', 'Please fix quiz errors before adding the lesson.');
+                $this->addError('newLesson.assessment', 'Please fix quiz errors before adding the lesson.');
                 $this->dispatch('swal', [
                     'title' => 'Invalid Assessment',
                     'html' => 'Please fix assessment errors before adding the lesson.',
@@ -77,12 +96,13 @@ class LessonCreate extends Component {
                 return;
             }
 
-            $this->dispatch('lesson-added', newLesson: $this->lesson, moduleIndex: $this->moduleIndex);
-            $this->dispatch('close-modal', id: 'addLesson');
 
-            $this->existingLessonTitles[] = $this->lesson['title'];
+            $this->dispatch('lesson-added', newLesson: $this->newLesson);
+            $this->dispatch('close-modal', id: 'addLessonModal');
 
-            $this->reset('lesson');
+            $this->existingLessonTitles[] = $this->newLesson['title'];
+
+            $this->reset('newLesson');
         } catch (Exception $e) {
             $this->dispatch('swal', [
                 'title' => 'Error',
@@ -93,17 +113,41 @@ class LessonCreate extends Component {
         }
     }
 
-    public function cancel(): void
+    #[On('assessment-updated')]
+    public function onAssessmentValid(bool $isValid): void
     {
-        if ($this->lesson['type'] === 'video') {
-            if (isset($this->lesson['tmp_video_file_name'])) {
-                File::delete(storage_path('app/private/livewire-tmp/' . $this->lesson['tmp_video_file_name']));
-            } else {
-                Storage::disk('public')->delete('course/videos/' . $this->lesson['video_file_name']);
-            }
+        $this->assessmentValid = $isValid;
+        if ($isValid) {
+            $this->resetErrorBag();
+            $this->resetValidation();
+        } else {
+            $this->addError('newLesson.assessment', 'Please fix quiz errors before adding the lesson.');
         }
-        $this->reset('lesson');
-        $this->dispatch('close-modal', id: 'addLesson');
+
+    }
+
+
+    #[On('video-saved')]
+    public function saveVideo(string $videoFileName, int $duration): void
+    {
+        $this->newLesson['video_file_name'] = $videoFileName;
+        $this->newLesson['duration'] = $duration;
+        if (isset($this->newLesson['tmp_video_file_name'])) {
+            unset($this->newLesson['tmp_video_file_name']);
+        }
+    }
+
+    #[On('tmp-video-uploaded')]
+    public function tmpVideoUpload(string $tmpVideoFileName): void
+    {
+        $this->newLesson['tmp_video_file_name'] = $tmpVideoFileName;
+    }
+
+    #[On('video-changed-or-deleted')]
+    public function changeOrDeletedVideo(): void
+    {
+        $this->newLesson['video_file_name'] = '';
+        $this->newLesson['duration'] = 0;
     }
 
     public function render(): Factory|Application|View
