+++ app/Services/CourseCreation/Builders/AssessmentTypes/ProgrammingService.php
@@ -2,7 +2,7 @@
 
 namespace App\Services\CourseCreation\Builders\AssessmentTypes;
 
-use App\Livewire\Client\CourseCreation\Components\Builders\AssessmentTypes\Programming;
+use App\Livewire\Client\CourseCreation\Components\Builders\Lesson\LessonTypes\AssessmentTypes\Programming;
 
 class ProgrammingService {
     protected array $typeMap;
