+++ app/View/Components/Client/Dashboard/Inputs/Select.php
@@ -8,9 +8,6 @@ use Illuminate\Support\Collection;
 use Illuminate\View\Component;
 
 class Select extends Component {
-    /**
-     * Create a new component instance.
-     */
     public function __construct(
         public string           $name,
         public string           $label,
@@ -19,11 +16,9 @@ class Select extends Component {
         public string $info,
         public bool             $isError = false,
         public string $model = '',
+        public bool   $isBoostrapSelect = true,
     ) {}
 
-    /**
-     * Get the view / contents that represent the component.
-     */
     public function render(): View|Closure|string
     {
         return view('components.client.dashboard.inputs.select');
