+++ resources/js/app.js
@@ -62,3 +62,15 @@ window.autocompletion = autocompletion;
 window.markdown = markdown;
 
 import '../assets/js/main.js';
+
+document.addEventListener('livewire:init', () => {
+    Livewire.on('close-modal', ({id}) => {
+        let modalEl = document.getElementById(id);
+        if (modalEl) {
+            let modal = bootstrap.Modal.getInstance(modalEl)
+                || new bootstrap.Modal(modalEl);
+            modal.hide();
+        }
+    });
+});
+
