--- resources/views/components/client/dashboard/course-creation/builders/assessment-types/quiz/option-editor.blade.php
@@ -1,3 +1,0 @@
-<div>
-    <!-- It is never too late to be what you might have been. - George Eliot -->
-</div>
