+++ resources/views/livewire/client/course-creation/components/builders/lesson/lesson-types/assessment-types/programming.blade.php
@@ -9,6 +9,7 @@
                 label="Problem Title"
                 placeholder="e.g., Two Sum, Valid Parentheses, Longest Substring"
                 info="Enter a concise, descriptive title for your coding problem."
+                :isError="$errors->has('programming.title')"
             />
 
             <x-client.dashboard.inputs.markdown-area
@@ -26,8 +27,7 @@
                     </button>
                 </div>
                 <div class="content">
-                    <button class="awe-btn" @click.prevent="$wire.validateStep1().then(ok => ok && (step = 2))">Next
-                    </button>
+                    <button class="awe-btn" @click="$wire.validateStep1().then(ok => ok && (step = 2))">Next</button>
                 </div>
             </div>
         </div>
@@ -39,6 +39,7 @@
                 label="Method Signature"
                 placeholder="e.g., twoSum, findMedian, longestSubstring"
                 info="Enter the main method name in camelCase format (e.g., twoSum, reverseList)."
+                :isError="$errors->has('problem.function_name')"
             />
 
             <x-client.dashboard.inputs.select
@@ -49,6 +50,7 @@
                 :options="$typeMap"
                 placeholder="Select a data type"
                 info="Select the expected return type of the method."
+                :isError="$errors->has('problem.return_type')"
             />
 
             <div class="course-field mb--20 col-lg-6">
@@ -179,6 +181,7 @@
                                     name="newTestCase.inputs.{{ $index }}.value"
                                     label="{{ $param['name'] }}"
                                     info="Enter the input value for parameter '{{ $param['name'] }}' as {{ $typeMap[$param['type']]['label'] }} format (e.g: {{ $typeMap[$param['type']]['example'] }})."
+                                    :isError="$errors->has('newTestCase.inputs.' . $index . '.value')"
                                 />
 
                             @endforeach
@@ -188,6 +191,7 @@
                                     name="newTestCase.output.value"
                                     label="Expected Output"
                                     info="Enter the expected output value as {{ $typeMap[$problem['return_type']]['label'] }} format (e.g: {{ $typeMap[$problem['return_type']]['example'] }})."
+                                    :isError="$errors->has('newTestCase.output.value')"
                                 />
                             @endif
                             <div class="col-2 p-0">
@@ -264,7 +268,7 @@
                 },
             }),
         ],
-        doc: @json($programming['description'] ?? ''),
+        doc: '',
         parent: document.getElementById('programming-description{{ !empty($unique) ? '-' . $unique : '' }}-editor')
     });
 </script>
