+++ resources/views/livewire/client/course-creation/components/builders/lesson/lesson-types/document.blade.php
@@ -1,8 +1,10 @@
 <x-client.dashboard.inputs.markdown-area
-    id="lesson-document"
+    id="new-lesson-document"
     label="Document"
+    :isError="$errors->has('document')"
     info="Markdown is supported. This is the document that will be displayed to the student."
-    name="document"/>
+    name="document"
+/>
 
 @script
 <script>
@@ -28,13 +30,10 @@
                     padding: "10px",
                     fontSize: "13px",
                 },
-                ".cm-content": {
-                    caretColor: "#000"
-                },
+                ".cm-content": {caretColor: "#000"},
             }),
         ],
-        doc: @json($document),
-        parent: document.getElementById('lesson-document-editor'),
+        parent: document.getElementById('new-lesson-document-editor'),
     });
 </script>
 @endscript
